Make the update badge the update control
Clicking it starts the update, so it is the sign and the action at once.
That is what lets it work on the Models page, whose card carries no
overflow menu — a waiting version was invisible on the page the user
actually sits on.

The "⋯" menu keeps only Uninstall: an update offered in two places is one
more than a card needs, and the chip is the one it gives a reason to look
at.
